% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{revbayes}
\alias{revbayes}
\title{revbayes}
\usage{
revbayes(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run revbayes
}
\details{
Note, no interaction possible.
}
\examples{
library(outsider)
revbayes <- module_import('revbayes', repo = 'dombennett/om..revbayes')
revbayes('--help')
}
